chore: refactor disk methods in mixin
